Amend state
