+++ src/App.js
@@ -1,6 +1,6 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import "./App.css";
-import { Route } from "react-router-dom";
+import { Route, withRouter } from "react-router-dom";
 import Navbar from "./component/navbar";
 import Footer from "./component/footer";
 import Login from "./component/login";
@@ -9,6 +9,7 @@ import Dashboard from "./component/dashboard";
 import Register from "./component/register";
 import withAuthChecker from "./custom-hooks/withAuthChecker";
 import axiosWithAuth from "./custom-hooks/axiosWithAuth";
+import axios from "axios";
 
 function App(props) {
   const [loginFormValues, setLoginFormValues] = useState({
@@ -18,7 +19,7 @@ function App(props) {
   const [user, setUser] = useState();
   const [error, setError] = useState(null);
 
-  useEffect(() => {
+  const axiosOnLogin = e => {
     axiosWithAuth()
       .get("https://split-the-bill-api.herokuapp.com/api/users/profile")
       .then(response => {
@@ -27,7 +28,31 @@ function App(props) {
       .catch(err => {
         setError(err.message);
       });
-  }, []);
+  };
+
+  const handleSubmit = e => {
+    e.preventDefault();
+    axios
+      .post("https://split-the-bill-api.herokuapp.com/api/auth/login", {
+        email: loginFormValues.email,
+        password: loginFormValues.password
+      })
+      .then(response => {
+        localStorage.setItem("token", response.data.token);
+        props.history.push("/dashboard");
+        axiosOnLogin();
+      })
+      .catch(error => {
+        setError(error);
+      });
+  };
+
+  const handleChange = event => {
+    setLoginFormValues({
+      ...loginFormValues,
+      [event.target.name]: event.target.value
+    });
+  };
 
   const allowAccess = e => {
     return <Dashboard user={user} {...props} />;
@@ -35,7 +60,10 @@ function App(props) {
 
   return (
     <div className="App">
-      <Route path="/" render={props => <Navbar {...props} />} />
+      <Route
+        path="/"
+        render={props => <Navbar {...props} setUser={setUser} />}
+      />
       <Route exact path="/" component={Home} />
       <Route
         exact
@@ -45,6 +73,10 @@ function App(props) {
             {...props}
             loginFormValues={loginFormValues}
             setLoginFormValues={setLoginFormValues}
+            user={user}
+            setUser={setUser}
+            handleChange={handleChange}
+            handleSubmit={handleSubmit}
           />
         )}
       />
@@ -59,4 +91,4 @@ function App(props) {
   );
 }
 
-export default App;
+export default withRouter(App);
