+++ src/component/login.js
@@ -3,42 +3,43 @@ import axios from "axios";
 import { Link } from "react-router-dom";
 
 function Login(props) {
-  const handleSubmit = e => {
-    e.preventDefault();
+  // const handleSubmit = e => {
+  //   e.preventDefault();
 
-    axios
-      .post("https://split-the-bill-api.herokuapp.com/api/auth/login", {
-        email: props.loginFormValues.email,
-        password: props.loginFormValues.password
-      })
-      .then(response => {
-        localStorage.setItem("token", response.data.token);
-        props.history.push("/dashboard");
-      })
-      .catch(error => {
-        console.log(error);
-      });
-  };
+  //   axios
+  //     .post("https://split-the-bill-api.herokuapp.com/api/auth/login", {
+  //       email: props.loginFormValues.email,
+  //       password: props.loginFormValues.password
+  //     })
+  //     .then(response => {
+  //       localStorage.setItem("token", response.data.token);
+  //       props.history.push("/dashboard");
 
-  const handleChange = event => {
-    props.setLoginFormValues({
-      ...props.loginFormValues,
-      [event.target.name]: event.target.value
-    });
-  };
+  //     })
+  //     .catch(error => {
+  //       console.log(error);
+  //     });
+  // };
+
+  // const handleChange = event => {
+  //   props.setLoginFormValues({
+  //     ...props.loginFormValues,
+  //     [event.target.name]: event.target.value
+  //   });
+  // };
 
   return (
     <div>
-      <form className="forms" onSubmit={handleSubmit}>
+      <form className="forms" onSubmit={props.handleSubmit}>
         <input
-          onChange={handleChange}
+          onChange={props.handleChange}
           name="email"
           type="text"
           value={props.loginFormValues.email}
           placeholder="Email"
         />
         <input
-          onChange={handleChange}
+          onChange={props.handleChange}
           name="password"
           type="password"
           value={props.loginFormValues.password}
